+++ src/Views/Filters/MultiSelectDropdownFilter.php
@@ -8,6 +8,8 @@ class MultiSelectDropdownFilter extends Filter
 {
     public array $options = [];
 
+    public string $viewPath = 'livewire-tables::components.tools.filters.multi-select-dropdown';
+
     protected string $firstOption = '';
 
     public function setFirstOption(string $firstOption): MultiSelectDropdownFilter
@@ -107,6 +109,6 @@ class MultiSelectDropdownFilter extends Filter
 
     public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
     {
-        return view('livewire-tables::components.tools.filters.multi-select-dropdown', $this->getFilterDisplayData());
+        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }