+++ src/Views/Filters/NumberFilter.php
@@ -6,6 +6,8 @@ use Rappasoft\LaravelLivewireTables\Views\Filter;
 
 class NumberFilter extends Filter
 {
+    public string $viewPath = 'livewire-tables::components.tools.filters.number';
+
     public function validate(mixed $value): int|bool
     {
         return is_numeric($value) ? $value : false;
@@ -26,6 +28,6 @@ class NumberFilter extends Filter
 
     public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
     {
-        return view('livewire-tables::components.tools.filters.number', $this->getFilterDisplayData());
+        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }