+++ src/Views/Filters/NumberRangeFilter.php
@@ -8,6 +8,8 @@ class NumberRangeFilter extends Filter
 {
     public array $options = [];
 
+    public string $viewPath = 'livewire-tables::components.tools.filters.number-range';
+
     public function options(array $options = []): NumberRangeFilter
     {
         $this->options = [...config('livewire-tables.numberRange.defaultOptions'), ...$options];
@@ -88,6 +90,6 @@ class NumberRangeFilter extends Filter
 
     public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
     {
-        return view('livewire-tables::components.tools.filters.number-range', $this->getFilterDisplayData());
+        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }