+++ src/Views/Filters/SelectFilter.php
@@ -8,6 +8,8 @@ class SelectFilter extends Filter
 {
     public array $options = [];
 
+    public string $viewPath = 'livewire-tables::components.tools.filters.select';
+
     public function options(array $options = []): SelectFilter
     {
         $this->options = $options;
@@ -63,6 +65,6 @@ class SelectFilter extends Filter
 
     public function render(): string|\Illuminate\Contracts\Foundation\Application|\Illuminate\View\View|\Illuminate\View\Factory
     {
-        return view('livewire-tables::components.tools.filters.select', $this->getFilterDisplayData());
+        return view($this->getViewPath(), $this->getFilterDisplayData());
     }
 }