+++ src/Views/Traits/Configuration/FilterConfiguration.php
@@ -139,4 +139,11 @@ trait FilterConfiguration
 
         return $this;
     }
+
+    public function setCustomView(string $customView): self
+    {
+        $this->viewPath = $customView;
+
+        return $this;
+    }
 }